pauls comments
